Embed getCharSize and getGridSize in Redraw.ts
This way they'll be automatically removed when the old renderer is
killed.
