+++ src/render/Redraw.ts
@@ -1,7 +1,30 @@
 import { NvimMode } from "../utils/configuration";
 import { page } from "../page/proxy";
-import { getCharSize, getGridSize, parseGuifont, toHexCss } from "../utils/utils";
+import { parseGuifont, toHexCss } from "../utils/utils";
 import { Grid } from "./Grid";
+
+// Returns a number tuple representing the size of characters in the host
+export function getCharSize(host: HTMLElement) {
+    const span = document.createElement("span");
+    span.style.position = "absolute";
+    span.style.top = "0px";
+    span.style.left = "0px";
+    span.innerText = " ";
+    host.appendChild(span);
+    const { width, height } = span.getBoundingClientRect();
+    host.removeChild(span);
+    return [width, height];
+}
+
+// Returns a number tuple representing how many columns and rows can fit in the
+// host.
+export function getGridSize(host: HTMLElement) {
+    const rect = host.getBoundingClientRect();
+    const [width, height] = getCharSize(host);
+    return [Math.floor(rect.width / width), Math.floor(rect.height / height)];
+}
+
+
 
 // Takes a `guifont` declaration and returns that same font declaration but as
 // a bunch of CSS declarations.
